Drive-Thru-Backend/Drive-Thru-Backend/requirements.txt
asgiref==3.5.0

Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
Pillow==9.0.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
